feat: major refactor
To reduce memory usage, the `Observable` class now extends the `Map` class instead of using composition. Note: This means you can't use the cjs bundle without polyfilling the `Map` class, because it's impossible to extend the native `Map` class without a `class` declaration.

In the same vein, the `ObserverSet` interface is now the `ObservedValue` class which extends the native `Set` class. The `Observable` class no longer has `add` and `remove` methods. Instead, you'll want to call `get` (which creates an `ObservedValue` on-the-fly if undefined) and add your observer directly using traditional `Set` methods (which `ObservedValue` inherits).

All `ObservedValue` instances live in memory for as long as their `Observable` owner does. This lets us allocate instances during render without needing to worry about cleanup. Besides, rarely does a value get observed just once, so there's really no need to aggressively deallocate unused instances. Even if the value is never observed again, the memory impact is negligible.

The `_onChange` abstract method of `Observer` is now a nullable property. It's nullified when the observer is disposed. Lastly, it's now called `onChange` (no underscore prefix).

The `Auto` class no longer extends the `Observer` class, because it needs the ability to track observable access without immediately observing the tracked observables. To do this, it creates an immutable `AutoObserver` instance that holds a nonce for every observed value. When the `Auto` instance is ready to observe them, call its `commit` method to replace the previous immutable observer. This allows for `Auto` to work in React's concurrent mode, because it no longer needs to subscribe during render. Look at the `withAuto` function in this diff to see the kind of usage `Auto` needs to support.

For clarity, the `run` method of `Auto` no longer registers an observer for any observable values accessed within the effect. You need to call `commit` for that now. If you call `commit` immediately after calling `run`, pass true to avoid an unnecessary dirty check.

The `when` function no longer leaks memory when it resolves or rejects its promise.

The `withAuto` function now checks for changes between render and commit, which triggers a sync re-render to ensure stale data is never displayed to the user.

Added an `onDelay` property to the `AutoConfig` object, which allows for custom scheduling of async reactions.