feat: add "createHook" export
This function takes a piece of state and returns a React hook for it.

When the hook is called, it returns the state, except wrapped with a proxy.

Accessing data within the proxied state is tracked. When any of the tracked values are changed, the component is re-rendered.

This hook should be compatible with React concurrent mode.