refactor: remove "createHook"
Not worth introducing the concept of a render-aware observable when "withAuto" works without any gotchas.

The hook-provided observable can leak to children very easily, resulting in less performant re-renders and more cognitive load
when deciding if that trade-off is worth it (which it usually never is). In the end, it causes more confusion than it helps.

In the future, "createHook" could have its own package where "withAuto" is replaced with another HOC factory which creates
a sort of render boundary for efficient re-rendering when a subtree's observed values are changed.