fix: add ObservableMap and ObservableSet
When the proxy doesn't inherit from the Map or Set class, it can't use methods intended *only* for those classes. Otherwise an error like "incompatible receiver" is thrown.

The ObservableMap and ObservableSet instances don't store any data. They need to inherit from Map and Set respectively so that "instanceof" works as expected.

The alternative to subclassing was binding Map and Set methods to the original object on every access.