fix: make assignments work again in JavaScriptCore
The JSC engine used by React Native has the following edge case.

When a proxied object has a property assigned, and the proxy handler has `defineProperty` defined and `set` is undefined, the descriptor that is received by the `defineProperty` handler has all of its attributes set to false. This means a basic assignment (eg: `obj.a = 1`) will define a non-writable and non-configurable property.