feat(useBinding): skip observation inside withAuto components
Manual observation isn't necessary when a `withAuto` component is being used. This commit is great for React hooks that want to use observable data for re-rendering purposes in any component (even ones that don't use withAuto), because it synchronizes with `withAuto` updates when possible, instead of always calling its own `forceUpdate` as it did before this commit.